Add post repository
Added post repository and fixed id validation bug
