+++ controllers/validators/id.js
@@ -1,3 +1,3 @@
 exports.validObjectId = function (id) {
-    return /[a-f0-9]{24}/.test(id) ? true : false;
+  return /^[a-f0-9]{24}$/.test(id) ? true : false;
 };
